+++ truffle/src/com.oracle.truffle.polyglot/src/com/oracle/truffle/polyglot/PolyglotEngineImpl.java
@@ -68,6 +68,7 @@ import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Objects;
 import java.util.Set;
+import java.util.concurrent.CancellationException;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
@@ -79,7 +80,6 @@ import java.util.function.Predicate;
 import java.util.function.Supplier;
 import java.util.logging.Level;
 
-import com.oracle.truffle.api.impl.DefaultTruffleRuntime;
 import org.graalvm.collections.EconomicSet;
 import org.graalvm.collections.Equivalence;
 import org.graalvm.home.HomeFinder;
@@ -111,6 +111,7 @@ import com.oracle.truffle.api.TruffleOptions;
 import com.oracle.truffle.api.dsl.Specialization;
 import com.oracle.truffle.api.dsl.SpecializationStatistics;
 import com.oracle.truffle.api.exception.AbstractTruffleException;
+import com.oracle.truffle.api.impl.DefaultTruffleRuntime;
 import com.oracle.truffle.api.impl.DispatchOutputStream;
 import com.oracle.truffle.api.instrumentation.ContextsListener;
 import com.oracle.truffle.api.instrumentation.ThreadsListener;
@@ -219,7 +220,7 @@ final class PolyglotEngineImpl implements com.oracle.truffle.polyglot.PolyglotIm
 
     final SpecializationStatistics specializationStatistics;
     Function<String, TruffleLogger> engineLoggerSupplier;   // effectively final
-    private volatile TruffleLogger engineLogger;
+    @CompilationFinal private TruffleLogger engineLogger;   // effectively final
 
     final WeakAssumedValue<PolyglotContextImpl> singleContextValue = new WeakAssumedValue<>("single context");
 
@@ -251,7 +252,7 @@ final class PolyglotEngineImpl implements com.oracle.truffle.polyglot.PolyglotIm
     PolyglotEngineImpl(PolyglotImpl impl, SandboxPolicy sandboxPolicy, String[] permittedLanguages,
                     DispatchOutputStream out, DispatchOutputStream err, InputStream in, OptionValuesImpl engineOptions,
                     Map<String, Level> logLevels,
-                    EngineLoggerProvider engineLogger, Map<String, String> options,
+                    EngineLoggerProvider engineLoggerSupplier, Map<String, String> options,
                     boolean allowExperimentalOptions, boolean boundEngine, boolean preInitialization,
                     MessageTransport messageInterceptor, LogHandler logHandler,
                     TruffleLanguage<Object> hostImpl, boolean hostLanguageOnly, AbstractPolyglotHostService polyglotHostService) {
@@ -296,7 +297,8 @@ final class PolyglotEngineImpl implements com.oracle.truffle.polyglot.PolyglotIm
                                 idToInstrument.get(id).cache.getClassName());
             }
         }
-        this.engineLoggerSupplier = engineLogger;
+        this.engineLoggerSupplier = engineLoggerSupplier;
+        this.engineLogger = initializeEngineLogger(engineLoggerSupplier, logLevels);
         this.engineOptionValues = engineOptions;
 
         Map<String, Object> publicLanguages = new LinkedHashMap<>();
@@ -342,7 +344,7 @@ final class PolyglotEngineImpl implements com.oracle.truffle.polyglot.PolyglotIm
             this.specializationStatistics = null;
         }
 
-        this.runtimeData = RUNTIME.createRuntimeData(this, engineOptions, engineLogger, sandboxPolicy);
+        this.runtimeData = RUNTIME.createRuntimeData(this, engineOptions, engineLoggerSupplier, sandboxPolicy);
         notifyCreated();
 
         if (!preInitialization) {
@@ -514,6 +516,7 @@ final class PolyglotEngineImpl implements com.oracle.truffle.polyglot.PolyglotIm
         this.probeAssertionsEnabled = prototype.probeAssertionsEnabled;
         this.hostLanguage = createLanguage(LanguageCache.createHostLanguageCache(prototype.getHostLanguageSPI()), HOST_LANGUAGE_INDEX, null);
         this.engineLoggerSupplier = prototype.engineLoggerSupplier;
+        this.engineLogger = prototype.engineLogger;
 
         this.polyglotHostService = prototype.polyglotHostService;
         this.internalResourceRoots = prototype.internalResourceRoots;
@@ -612,20 +615,15 @@ final class PolyglotEngineImpl implements com.oracle.truffle.polyglot.PolyglotIm
     }
 
     TruffleLogger getEngineLogger() {
-        TruffleLogger result = this.engineLogger;
-        if (result == null) {
-            synchronized (this.lock) {
-                result = this.engineLogger;
-                if (result == null) {
-                    result = this.engineLoggerSupplier.apply(OPTION_GROUP_ENGINE);
-                    Object logger = EngineAccessor.LANGUAGE.getLoggerCache(result);
-                    LoggerCache loggerCache = (LoggerCache) EngineAccessor.LANGUAGE.getLoggersSPI(logger);
-                    loggerCache.setOwner(this);
-                    EngineAccessor.LANGUAGE.configureLoggers(this, logLevels, logger);
-                    this.engineLogger = result;
-                }
-            }
-        }
+        return engineLogger;
+    }
+
+    private TruffleLogger initializeEngineLogger(Function<String, TruffleLogger> supplier, Map<String, Level> levels) {
+        TruffleLogger result = supplier.apply(OPTION_GROUP_ENGINE);
+        Object logger = EngineAccessor.LANGUAGE.getLoggerCache(result);
+        LoggerCache loggerCache = (LoggerCache) EngineAccessor.LANGUAGE.getLoggersSPI(logger);
+        loggerCache.setOwner(this);
+        EngineAccessor.LANGUAGE.configureLoggers(this, levels, logger);
         return result;
     }
 
@@ -693,7 +691,7 @@ final class PolyglotEngineImpl implements com.oracle.truffle.polyglot.PolyglotIm
          */
         this.storeEngine = this.storeEngine || RUNTIME.isStoreEnabled(engineOptions);
         this.engineLoggerSupplier = logSupplier;
-        this.engineLogger = null;
+        this.engineLogger = initializeEngineLogger(logSupplier, newLogConfig.logLevels);
         /*
          * The engineLoggers must be created before calling PolyglotContextImpl#patch, otherwise the
          * engineLoggers is not in an array returned by the PolyglotContextImpl#getAllLoggers and
@@ -1559,6 +1557,8 @@ final class PolyglotEngineImpl implements com.oracle.truffle.polyglot.PolyglotIm
                     }
                 } catch (ExecutionException | InterruptedException e) {
                     throw CompilerDirectives.shouldNotReachHere(e);
+                } catch (CancellationException e) {
+                    // Expected
                 }
                 if (timedOut) {
                     return false;
@@ -2375,7 +2375,7 @@ final class PolyglotEngineImpl implements com.oracle.truffle.polyglot.PolyglotIm
 
     @SuppressWarnings("try")
     void validateVirtualThreadCreation() {
-        if (!warnedVirtualThreadSupport.compareAndExchange(false, true)) {
+        if (!warnedVirtualThreadSupport.get() && warnedVirtualThreadSupport.compareAndSet(false, true)) {
             try (AbstractPolyglotImpl.ThreadScope scope = impl.getRootImpl().createThreadScope()) {
                 var options = getEngineOptionValues();
                 boolean warnVirtualThreadSupport = options.get(PolyglotEngineOptions.WarnVirtualThreadSupport);
@@ -2384,8 +2384,7 @@ final class PolyglotEngineImpl implements com.oracle.truffle.polyglot.PolyglotIm
                     if (!TruffleOptions.AOT) {
                         getEngineLogger().warning("""
                                         Using polyglot contexts on Java virtual threads on HotSpot is experimental in this release,
-                                        because access to caller frames in write or materialize mode is not yet supported on virtual threads (some tools and languages depend on that),
-                                        and there is a limit of maximum 65535 virtual threads concurrently accessing the context at the same time.
+                                        because access to caller frames in write or materialize mode is not yet supported on virtual threads (some tools and languages depend on that).
                                         To disable this warning use the '--engine.WarnVirtualThreadSupport=false' option or the '-Dpolyglot.engine.WarnVirtualThreadSupport=false' system property.
                                         """);
                     } else {